feat: added ACL in orders, categories, users

+++ controllers/categoriesController.js
@@ -1,5 +1,7 @@
 const Category = require('../models/categories');
 const Errors = require('../models/Errors');
+const Roles = require('../auth/acl').Roles;
+const createError = require('http-errors');
 
 class categoriesController {
     static async index (req, res, next) {
@@ -31,54 +33,70 @@ class categoriesController {
     }
 
     static async readAttributtes (req, res, next) {
-        let category = await new Category().getAttributesByCategory(req.params.id);
-        if(category==undefined) {
-            next(Errors('404'))
-        } else if (category.status) {
-            next(category);
+        if (req.user.role  == Roles.ADMIN) {
+            let category = await new Category().getAttributesByCategory(req.params.id);
+            if (category == undefined) {
+                next(Errors('404'))
+            } else if (category.status) {
+                next(category);
+            } else {
+                res.status(200).json({
+                    data: category,
+                    message: "Атрибуты для данной категории получены успешно",
+                    responseCode: 0,
+                })
+            }
         } else {
-            res.status(200).json({
-                data: category,
-                message: "Атрибуты для данной категории получены успешно",
-                responseCode: 0,
-            })
+            next(createError(403, "Не хватает прав"));
         }
     }
 
     static async write (req, res, next) {
-        let category = await new Category().createCategory(req.body);
-        if (category.status) {
-            next(category);
+        if (req.user.role  == Roles.ADMIN) {
+            let category = await new Category().createCategory(req.body);
+            if (category.status) {
+                next(category);
+            } else {
+                res.status(200).json({
+                    data: category,
+                    message: `Новая категория успешно создана`,
+                    responseCode: 0,
+                })
+            }
         } else {
-            res.status(200).json({
-                data: category,
-                message: `Новая категория успешно создана`,
-                responseCode: 0,
-            })
+            next(createError(403, "Не хватает прав"));
         }
     }
 
     static async update (req, res, next) {
-        let category = await new Category().updateCategory(req.body);
-        if (category.status) {
-            next(category);
+        if (req.user.role  == Roles.ADMIN) {
+            let category = await new Category().updateCategory(req.body);
+            if (category.status) {
+                next(category);
+            } else {
+                res.status(200).json({
+                    message: `Данные категории - ${category} обновлены`,
+                    responseCode: 0,
+                })
+            }
         } else {
-            res.status(200).json({
-                message: `Данные категории - ${category} обновлены`,
-                responseCode: 0,
-            })
+            next(createError(403, "Не хватает прав"));
         }
     }
 
     static async delete (req, res, next) {
-        let category = await new Category().deleteCategory(req.body.id);
-        if (category.status) {
-            next(category);
+        if (req.user.role  == Roles.ADMIN) {
+            let category = await new Category().deleteCategory(req.body.id);
+            if (category.status) {
+                next(category);
+            } else {
+                res.status(200).json({
+                    message: "Категория успешно удалена",
+                    responseCode: 0,
+                })
+            }
         } else {
-            res.status(200).json({
-                message: "Категория успешно удалена",
-                responseCode: 0,
-            })
+            next(createError(403, "Не хватает прав"));
         }
     }
 }
