+++ controllers/ordersController.js
@@ -1,9 +1,11 @@
 const Order = require('../models/orders');
 const Errors = require('../models/Errors');
+const Roles = require('../auth/acl').Roles;
+const createError = require('http-errors');
 
 class ordersController {
     static async index (req, res, next) {
-        let order = await new Order().getOrdersByUser(req.body['user_id']);
+        let order = await new Order().getOrdersByUser(req.user.id);
         if (order.status && Number.isInteger(order.status)) {
             next(order);
         } else {
@@ -16,8 +18,8 @@ class ordersController {
     }
 
     static async read (req, res, next) {
-        let order = await new Order().getOrderWithProducts(req.params.id);
-        if(order==undefined) {
+        let order = await new Order().getOrderWithProducts(req.params.id, req.user.id);
+        if (order == undefined) {
             next(Errors('404'))
         } else if (order.status && Number.isInteger(order.status)) {
             next(order);
@@ -31,7 +33,7 @@ class ordersController {
     }
 
     static async write (req, res, next) {
-        let order = await new Order().createOrderWithProducts(req.body);
+        let order = await new Order().createOrderWithProducts(req.body, req.user.id);
         if (order.status && Number.isInteger(order.status)) {
             next(order);
         } else {
@@ -43,26 +45,34 @@ class ordersController {
     }
 
     static async update (req, res, next) {
-        let order = await new Order().updateOrder(req.body);
-        if (order.status && Number.isInteger(order.status)) {
-            next(order);
+        if (req.user.role  == Roles.ADMIN) {
+            let order = await new Order().updateOrder(req.body);
+            if (order.status && Number.isInteger(order.status)) {
+                next(order);
+            } else {
+                res.status(200).json({
+                    message: `Данные заказа №${order} успешно изменены`,
+                    responseCode: 0,
+                })
+            }
         } else {
-            res.status(200).json({
-                message: `Данные заказа №${order} успешно изменены`,
-                responseCode: 0,
-            })
+            next(createError(403, "Не хватает прав"));
         }
     }
 
     static async delete (req, res, next) {
-        let order = await new Order().deleteOrder(req.body.id);
-        if (order.status && Number.isInteger(order.status)) {
-            next(order);
+        if (req.user.role  == Roles.ADMIN) {
+            let order = await new Order().deleteOrder(req.body.id);
+            if (order.status && Number.isInteger(order.status)) {
+                next(order);
+            } else {
+                res.status(200).json({
+                    message: `Заказ №${order} успешно удален`,
+                    responseCode: 0,
+                })
+            }
         } else {
-            res.status(200).json({
-                message: `Заказ №${order} успешно удален`,
-                responseCode: 0,
-            })
+            next(createError(403, "Не хватает прав"));
         }
     }
 }
